pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "invoicing"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "nerc-rates>=1.0.1,<2.0.0",
    "pandas>=3.0,<4.0",
    "pyarrow",
    "boto3>=1.42.6,<2.0",
    "Jinja2",
    "validators",
    "python-dateutil",
    "pydantic>=2.0",
    "pydantic-settings",
    "PyYAML",
]
